pyproject.toml
[project]
name = "swarmllm"
version = "0.1.0"
description = "Multi-Agent LLM Swarm for Optimization"
requires-python = ">=3.10"
dependencies = [
    "gurobipy>=13.0.1",
    "mlx-lm>=0.31.1",
    "networkx>=3.4.2",
    "numpy>=1.24.0",
    "psutil>=7.2.2",
    "pydantic>=2",
    "pydantic-ai-slim[openai]>=1.73.0",
    "rich>=14.3.3",
    "scipy>=1.15.3",
    "tomli>=2.4.1",
]

[project.scripts]
swarmllm = "scripts.run:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
